New helper for chart functions
